soem api cleanup
